requirements.txt
dill==0.4.0
fastapi==0.115.12
langchain==0.3.23
langchain_chroma==0.2.3
langchain_community==0.3.21
langchain_core==0.3.54
langchain_google_genai==2.1.3
langchain_huggingface==0.1.2
nltk==3.9.1
pandas==2.2.3
pydantic==2.11.3
pypdf==5.4.0
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
streamlit==1.44.1
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
uvicorn==0.34.2
